app_satellite_image_processing/src/image_processing/requirements.txt
pip~=24.0
pillow~=10.3.0
tornado~=6.4
Jinja2~=3.1.4
python-dotenv~=1.0.1
opencv-python~=4.10.0.84
supervision~=0.24.0
roboflow~=1.1.48
numpy~=1.26.4
psycopg2-binary~=2.9.6